batch: structures, block roster and model meshing + five review cleanups
Everything that had accumulated in the working tree, in one commit.

## Structures (STRUCT)

Structure storage, placement and loot: an NBT reader, structure templates with
rotation, the template manager and placer, jigsaw pools and their expansion, the
structure placement/generation pass, and chest loot tables filled at generation
time (GameRuntime::createStructureChests). Templates and loot come from the
resource/data pack the player supplies -- no Mojang content enters the tree.

## Blocks, models and world generation

- `enum class Block` widened u8 -> u16 and the roster grown to ~196 blocks (the
  old 140 hit the 256 ceiling). Blocks serialize by name, so no save format bump.
- `BlockState::rawId` widened u16 -> u32 so the interned block+properties state
  space stops running out. The id is never written to disk or the wire (a section
  stores a local palette index, saves store name+properties), so old worlds load
  unchanged.
- General block-model meshing: FaceBakery + ElementModelBaker mesh a per-element
  vanilla model instead of a full cube, which is what the diodes and the lever
  actually are.
- Worldgen: a deepslate band below y=0 with a randomised gradient, deepslate ore
  variants and copper ore, applied after carving so caves and ores survive.
- y<0 consistency: kVoidDespawnY (kMinY - 64) replaces the hardcoded -8/0/256
  void checks, so mining below y=-8 no longer deletes the drop before it can be
  picked up.
- Translucent blocks render double-sided (stained glass against a solid block was
  showing through, because the translucent pipeline reused the opaque pass's
  back-face culling).

## Code-review cleanups

Five issues raised in docs/CODE_PROBLEMS.md, checked and fixed:

1. Retire the bundled vanilla resource line. DirectoryResourceProvider mapped
   `resources/vanilla/1.16.1/...` with three category renames, filled by an
   extraction script run against a local Minecraft install -- a tree that has not
   existed since the switch to standard resource packs, so every texture, sound,
   font and lang lookup missed and fell through to the pack. It now serves this
   project's own assets only and returns an empty path for anything Mojang-shaped;
   the startup "extracted Minecraft 1.16.1 resources were not found" warning and
   the misleading resource-root print are gone, and so is the extraction script.
   tools/entity_uv_lib.py reads entity skins from the installed packs instead.

2. The species manifest is the only way to define a creature. Pig, Cow and Zombie
   were hand-written classes with their own `::type()` accessors, kept alive only
   because callers held those accessors, while every newer species was a manifest
   row. They are manifest rows now, in their original registration order so the
   dense network ids are unchanged (pig 0, cow 1, zombie 2, chicken 3, sheep 4,
   husk 5). Callers name a species with `builtinSpecies("cow")`; spawn-egg
   suppliers are one template instantiated on the name. Also drops the redundant
   registerBuiltinEntities() call the renderer made after Application had already
   run it.

3. The smoke-test script leaves the render loop. 175 lines of
   `else if (smokeTest && frame == N)` interleaved with the frame's real work,
   plus five wait-state locals carried across every frame of every run, become a
   scheduler (render/SmokeScript.hpp) and a step list
   (render/vulkan/SmokeScriptSteps.hpp). run() drops from 955 to ~590 lines and a
   release run builds no script at all.

4. One definition per cycling option. Each option was stated three times -- a
   hand-rolled step in a cb.cycleXxx lambda, a formatting case in widgetLabel(),
   and the field itself. ui/OptionCycle.hpp is now a constexpr table: values,
   GameOptions member pointer and translation keys per row, with a generic
   `cycleOptionValue(desc, options, +/-1)` (backwards works, so a two-way selector
   needs no new plumbing) and a generic label. The renderer keeps only
   applyOptionChanged(), the single place a change's side effect lives.

5. One modal precedence for input. The five GLFW callbacks each opened with their
   own chain of `if (capturing) ... if (chatOpen) ... if (paused) ...`, in a
   different order each, and the per-frame poll gate spelled the same rule out
   twice more. input/ScreenMode.hpp states it once as a pure function
   (KeyCapture > TextField > Chat > Inventory > Menu > Play); the callbacks switch
   on it and the poll gate derives from it. The callback block drops from 239 to
   112 lines. The equivalence test found the one state the old copies disagreed
   with each other about (inventory open while paused, unreachable since
   setPaused closes the inventory) and it is now pinned.

Build fix: the Windows cross-build passes -Wa,-mbig-obj. VulkanRenderer.cpp had
been sitting just under COFF's 32767-section cap and crossed it.

New headless tests: smoke_script, option_cycle, screen_mode (the last one
brute-forces every flag combination against the expressions it replaced). Linux
debug builds clean and ctest is 202/202.

+++ src/world/BlockState.hpp
@@ -256,7 +256,7 @@ class BlockState final {
     // The raw id, for the section palette and for hashing. Not stable across
     // builds — the save layer must map it through identifiers and property
     // names, exactly as it already does for Block.
-    [[nodiscard]] constexpr std::uint16_t rawId() const { return id_; }
+    [[nodiscard]] constexpr std::uint32_t rawId() const { return id_; }
     // Rebuilds a state from a raw id — a chunk section's palette entry, or the
     // save layer round-tripping a cell. An id at or above the built-in table
     // (isUnknownStateId) names an UnknownBlock placeholder: content this build's
@@ -264,7 +264,7 @@ class BlockState final {
     // name and properties back. Such an id is preserved rather than clamped here;
     // every metadata accessor already clamps it to block 0's air-like defaults
     // (see validatedBlockStateId), so the cell stays inert without a special case.
-    [[nodiscard]] static constexpr BlockState fromRawId(std::uint16_t value) {
+    [[nodiscard]] static constexpr BlockState fromRawId(std::uint32_t value) {
         BlockState state;
         state.id_ = value;
         return state;
@@ -282,7 +282,7 @@ class BlockState final {
         return static_cast<std::uint8_t>(requested > maximum ? maximum : requested);
     }
 
-    std::uint16_t id_ = 0U;
+    std::uint32_t id_ = 0U;
 };
 
 // The collision/selection shape a state carries lives in BlockShape.hpp
@@ -306,7 +306,13 @@ class BlockState final {
     return skyLightOpacity(state.block());
 }
 
-static_assert(sizeof(BlockState) == sizeof(std::uint16_t));
+// BlockState is exactly its raw id: a uint32 now (widened from uint16 so the
+// interned block-state space can grow past 65536 — see BlockStateTable.hpp). A
+// chunk cell never stores this directly; the section palette does, and each cell
+// holds only a small local index (ChunkSection), so the width change does not
+// grow world memory. The id is never serialised raw — saves and the wire carry
+// the block name + property names/values.
+static_assert(sizeof(BlockState) == sizeof(std::uint32_t));
 static_assert(BlockState{}.block() == Block::Air);
 // The schema's arithmetic, pinned at compile time: a crop's age is its own
 // property, and it no longer travels in the direction enum.
